refactor: seperate controllers and routes to new files
