+++ app.js
@@ -3,7 +3,7 @@ import { fileURLToPath } from 'node:url';
 import express from 'express';
 import morgan from 'morgan';
 import fs from 'fs';
-
+import tourRouter from './routes/tourRoutes.js';
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const app = express();
 const port = process.env.PORT || 3000;
@@ -15,89 +15,9 @@ const tours = JSON.parse(
   fs.readFileSync(`${__dirname}/dev-data/data/tours-simple.json`, 'utf-8')
 );
 
-const getAllTours = (req, res) => {
-  res.status(200).json({
-    status: 'Success',
-    results: tours.length,
-    data: {
-      tours,
-    },
-  });
-};
-const deleteTour = (req, res) => {
-  const { id } = req.params;
-  const tour = tours.find((el) => +el.id === +id);
-  if (!tour) {
-    return res.status(404).json({
-      status: 'Fail',
-      message: 'Invalid ID',
-    });
-  }
-  res.status(204).json({
-    status: 'Success',
-    data: null,
-  });
-};
-const createNewTour = async (req, res) => {
-  const newId = tours[tours.length - 1].id + 1;
-  const newTour = Object.assign({ id: newId }, req.body);
-  tours.push(newTour);
-  fs.writeFile(
-    `${__dirname}/dev-data/data/tours-simple.json`,
-    JSON.stringify(tours),
-    (err) => {
-      res.status(201).json({
-        status: 'Success',
-        data: {
-          newTour,
-        },
-      });
-    }
-  );
-};
-const getSingleTour = (req, res) => {
-  const { id } = req.params;
-  const tour = tours.find((el) => +el.id === +id);
-  if (!tour) {
-    return res.status(404).json({
-      status: 'Fail',
-      message: 'Invalid ID',
-    });
-  }
 
-  res.status(200).json({
-    status: 'Success',
-    data: {
-      tour,
-    },
-  });
-};
-const updateTour = (req, res) => {
-  const { id } = req.params;
-  let tour = tours.find((el) => +el.id === +id);
-  if (!tour) {
-    return res.status(404).json({
-      status: 'Fail',
-      message: 'Invalid ID',
-    });
-  }
-  tour = { ...tour, ...req.body };
-  tours[id] = tour;
-  res.status(201).json({
-    status: 'Success',
-    data: {
-      tour,
-    },
-  });
-};
 
-app.route('/api/v1/tours').get(getAllTours).post(createNewTour);
-
-app
-  .route('/api/v1/tours/:id')
-  .get(getSingleTour)
-  .patch(updateTour)
-  .delete(deleteTour);
+app.use('/api/v1/tours', tourRouter);
 
 app.listen(port, () => {
   console.log(`Application started running on port ${port}`);
